feature(ui-ux): move themedtext component inside commons folder
