+++ src/components/Header.tsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 import Link from "next/link";
 import { useState } from "react";
 import { truncateTextFromMiddle } from "@utils/textHelper";
-import ThemedText from "./ThemedText";
+import ThemedText from "./commons/ThemedText";
 import MetaMaskIcon from "./icons/MetaMaskIcon";
 
 interface Wallet {
